requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
numpy==1.23.1
opencv-python==4.6.0.66
aiohttp==3.8.1
python-dotenv==0.20.0
Pillow==9.2.0